+++ CHANGELOG.md
@@ -2,7 +2,7 @@ This is a history of changes to k13labs/futurama
 
 # 1.2.0
 * remove async-{future/channel/promise/...} variants of the `async` macro, replace uses with `async`.
-* added `thread`, `io-thread`, `compute-thread` macros to route work to the appropriate thread pool, such as :io, :compute, or :mixed.
+* added `thread`, updated `async`, both macros route work to the appropriate thread pool, such as :io, :compute, or :mixed.
 * added `*thread-factory*` dynamic binding to allow separately defining a factory-fn for `thread` calls, distinct from `async` calls.
 * deprecate the `completable-future` macro, uses of completable-future should be replaced with `thread`.
 * deprecate the `fixed-threadpool` function, instead prefer to use Executors thread pools according to need.